third_party/WebKit/LayoutTests/external/wpt/cookies/prefix/__secure.http.secure-expected.txt

This is a testharness.js-based test.
PASS __Secure: secure origin: 'Path=/;'
FAIL __Secure: secure origin: 'Secure;Path=/;' assert_equals: expected (string) "0.13232969518371385" but got (undefined) undefined
PASS __Secure: secure origin: 'Path=/;domain=not-web-platform.test'
FAIL __Secure: secure origin: 'Secure;Path=/;domain=not-web-platform.test' assert_equals: expected (string) "0.8627187387844981" but got (undefined) undefined
PASS __Secure: secure origin: 'Path=/;MaxAge=10'
FAIL __Secure: secure origin: 'Secure;Path=/;MaxAge=10' assert_equals: expected (string) "0.9587977318025727" but got (undefined) undefined
PASS __Secure: secure origin: 'Path=/;HttpOnly'
FAIL __Secure: secure origin: 'Secure;Path=/;HttpOnly' assert_equals: expected (string) "0.23905638776137872" but got (undefined) undefined
Harness: the test ran to completion.
